requirements.txt
alembic==1.8.1
aniso8601==9.0.1
apispec==6.0.2
apispec-webframeworks==0.5.2
attrs==22.1.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
exceptiongroup==1.0.4
flake8==6.0.0
flasgger==0.9.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
mccabe==0.7.0
mistune==2.0.4
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.10.0
